functional setting for cpt featured image
at this point, a featured image can be set and retrieved for the cpt
archive. creates a new setting, though, not sure if this is desirable.
could we just add on to the existing displayfeaturedimage setting, since
it is already serialized?

also, at this point, no validation is happening for the cpt images.