Set up site to use new WordPress image sizes
WordPress is adding new image sizes in version 5.3, so our custom image
size is redundant. Keeping it in for back compat, plus older images
will still need to be retrieved using the old image size name, or else
the original size will be used in lieu of a correctly sized option.